+++ src/main/java/kitchenpos/products/domain/tobe/Product.java
@@ -51,29 +51,15 @@ public class Product {
         return this.price.equals(price);
     }
 
+    public BigDecimal calculateSum(BigDecimal quantity) {
+        return price.getPrice().multiply(quantity);
+    }
+
     public UUID getId() {
         return id;
     }
 
     public BigDecimal getPrice() {
         return price.getPrice();
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        Product product = (Product) o;
-        return Objects.equals(id, product.id) && Objects.equals(name, product.name)
-                && Objects.equals(price, product.price);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, name, price);
     }
 }
