+++ src/test/java/kitchenpos/products/domain/tobe/ProductTest.java
@@ -53,7 +53,8 @@ class ProductTest {
     void Product의_가격을_마이너스로_변경하면_예외를_던진다() {
         Product product = new Product("치킨", BigDecimal.valueOf(20_000L));
 
-        assertThatThrownBy(() -> product.changePrice(new ProductPrice(BigDecimal.valueOf(-20_000L))))
+        assertThatThrownBy(
+                () -> product.changePrice(new ProductPrice(BigDecimal.valueOf(-20_000L))))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 }
